+++ queryprotocol/hermes2/protocol.go
@@ -56,9 +56,9 @@ type VoterInfo struct {
 
 // Ratio defines delegate reward distribution ratio
 type Ratio struct {
-	BlockRewardRatio     int
-	EpochRewardRatio     int
-	FoundationBonusRatio int
+	BlockRewardRatio     float64
+	EpochRewardRatio     float64
+	FoundationBonusRatio float64
 	EpochNumber          int
 }
